requirements.txt
fastapi
uvicorn
Jinja2
python-multipart
langchain
langgraph
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-dotenv